Drop support for PHP `5.6` and `7.x`
