+++ .github/workflows/test.yaml
@@ -8,11 +8,11 @@ jobs:
       fail-fast: false
       matrix:
         php-versions:
-          - '5.6'
-          - '7.4'
           - '8.0'
           - '8.1'
           - '8.2'
+          - '8.3'
+          - '8.4'
     steps:
       - name: Checkout
         uses: actions/checkout@v4
