+++ src/Log/LoggerDecorator.php
@@ -16,28 +16,6 @@ use Monolog\Handler\TestHandler;
 use Psr\Log\LoggerInterface;
 
 /**
- * Backwards compatibility with PSR Log v1.
- */
-$r = new \ReflectionClass(LoggerInterface::class);
-if (PHP_MAJOR_VERSION < 7 || !$r->getMethod('log')->hasReturnType()) {
-    /**
-     * Logger decorator trait that exposes PSR-3 Logger methods.
-     */
-    trait LoggerDecoratorTrait
-    {
-        use LoggerDecoratorV1Trait;
-    }
-} else {
-    /**
-     * Logger decorator trait that exposes PSR-3 Logger methods.
-     */
-    trait LoggerDecoratorTrait
-    {
-        use LoggerDecoratorV3Trait;
-    }
-}
-
-/**
  * Logger decorator that instantiates logger by configuration and exposes PSR-3 Logger methods
  */
 class LoggerDecorator implements LoggerInterface
@@ -58,7 +36,7 @@ class LoggerDecorator implements LoggerInterface
      *
      * @param LoggingConfig|null $config
      */
-    public function __construct(LoggingConfig $config = null)
+    public function __construct(?LoggingConfig $config = null)
     {
         $this->config = $config;
     }
