+++ src/Log/LoggerDecoratorTrait.php
@@ -13,21 +13,20 @@ namespace Cloudinary\Log;
 use UnexpectedValueException;
 
 /**
- * Logger decorator that instantiates logger by configuration and exposes PSR-3 v1 Logger methods
+ * Logger decorator that instantiates logger by configuration and exposes PSR-3 v3 Logger methods
  */
-trait LoggerDecoratorV1Trait
+trait LoggerDecoratorTrait
 {
     /**
      * Adds a log record at the DEBUG level.
      *
      * Used for logging fine-grained informational events that are most useful to debug an application.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function debug($message, array $context = [])
+    public function debug(string|\Stringable $message, array $context = []): void
     {
         $this->log('debug', $message, $context);
     }
@@ -35,13 +34,12 @@ trait LoggerDecoratorV1Trait
     /**
      * Logs with an arbitrary level
      *
-     * @param string $level   The log level
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string             $level   The log level
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function log($level, $message, array $context = [])
+    public function log($level, string|\Stringable $message, array $context = []): void
     {
         if ($this->logger === null) {
             $this->logger = LoggersList::instance()->getLogger($this->config);
@@ -68,12 +66,11 @@ trait LoggerDecoratorV1Trait
      *
      * Used for logging informational messages that highlight the progress of the application at coarse-grained level.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function info($message, array $context = [])
+    public function info(string|\Stringable $message, array $context = []): void
     {
         $this->log('info', $message, $context);
     }
@@ -83,12 +80,11 @@ trait LoggerDecoratorV1Trait
      *
      * Used for logging normal but significant events.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function notice($message, array $context = [])
+    public function notice(string|\Stringable $message, array $context = []): void
     {
         $this->log('notice', $message, $context);
     }
@@ -99,12 +95,11 @@ trait LoggerDecoratorV1Trait
      * Used for logging exceptional occurrences that are not errors.
      * Example: Use of deprecated APIs, poor use of an API, undesirable things that are not necessarily wrong.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function warning($message, array $context = [])
+    public function warning(string|\Stringable $message, array $context = []): void
     {
         $this->log('warning', $message, $context);
     }
@@ -114,12 +109,11 @@ trait LoggerDecoratorV1Trait
      *
      * Used for logging error events that might still allow the application to continue running.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function error($message, array $context = [])
+    public function error(string|\Stringable $message, array $context = []): void
     {
         $this->log('error', $message, $context);
     }
@@ -130,12 +124,11 @@ trait LoggerDecoratorV1Trait
      * Used for logging critical events.
      * Example: Application component unavailable, unexpected exception.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function critical($message, array $context = [])
+    public function critical(string|\Stringable $message, array $context = []): void
     {
         $this->log('critical', $message, $context);
     }
@@ -146,12 +139,11 @@ trait LoggerDecoratorV1Trait
      * Used for logging and probably triggering an alert that will wake someone up.
      * Example: Entire website down, database unavailable, etc.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function alert($message, array $context = [])
+    public function alert(string|\Stringable $message, array $context = []): void
     {
         $this->log('alert', $message, $context);
     }
@@ -161,12 +153,11 @@ trait LoggerDecoratorV1Trait
      *
      * Used for logging the highest level of errors. System is unusable.
      *
-     * @param string $message The log message
-     * @param array  $context The log context
+     * @param string|\Stringable $message The log message
+     * @param array              $context The log context
      *
-     * @return void
      */
-    public function emergency($message, array $context = [])
+    public function emergency(string|\Stringable $message, array $context = []): void
     {
         $this->log('emergency', $message, $context);
     }
